+++ CHANGELOG.md
@@ -627,3 +627,4 @@
 * Define `Betwixt` type family to avoid specifying intermediate categories
 * Define experimental `overlook` method
 * Remove `<$$$$>`, `<$$$$||=` and `=||$$$$>` operators
+* Define `Effectful` typeclass to delegate instance resolution for `Adaptable`